new save system alpha ver

+++ Assets/Scripts/UI/Menu/Game Intro/UICharacterPicker.cs
@@ -9,12 +9,12 @@ public class UICharacterPicker : MonoBehaviour, ISelectHandler
     public Image chracterPreview;
     public AudioClip selectedSound;
 
-    public delegate void OnClick(float characterIndex);
+    public delegate void OnClick(int characterIndex);
     public event OnClick onClick;
-    public delegate void OnHover(float characterIndex);
+    public delegate void OnHover(int characterIndex);
     public event OnHover onHover;
 
-    private float chraracterIndex = 0;
+    private int chraracterIndex = 0;
     public UICharacterPicker Load(int chraracterIndex)
     {
         this.chraracterIndex = chraracterIndex;
